Defined laserscanner's maximum viewing angle.
Calculate angle to cut off on either side of scanner to limit viewing angle to 180 deg
Changed minimum distance to stop to 0.2m before obstacle.
When the robot is in turn mode he will adapt the turn speed depending on the position of the obstacle. Obstacles infront means quicker turning.
Its a linear change of turning speed from 0 to 100% of maximum turning speed dependent on the angle.
There is a bug when the robot will get stuck in a position and will not turn out of it. Needs to be fixed